+++ src/eth/useTokenBalance.js
@@ -1,9 +1,6 @@
 import React, {useEffect, useState} from 'react';
-import { t } from 'saihubot-cli-adapter/dist/i18n';
 
 import {rpcTokenBalance} from '../helpers/ethRpc';
-import {i18nBalance} from '../i18n';
-import {formatAddress} from '../utils';
 
 export const useTokenBalance = ({
   addresses,
@@ -26,10 +23,10 @@ export const useTokenBalance = ({
           const tokenJson = await networkFetch(fetch, rpcTokenBalance(addresses[i], currentTokenAddr));
           if (tokenJson.result !== '0x0000000000000000000000000000000000000000000000000000000000000000') {
             data.push({
-              [t('addr', {i18n: i18nBalance})]: formatAddress(addresses[i]),
-              [t('token', {i18n: i18nBalance})]: tokenInfo.symbol,
-              [t('balance', {i18n: i18nBalance})]: Number(tokenJson.result) / 10 ** tokenInfo.decimals,
-              [t('source', {i18n: i18nBalance})]: tokenInfo.name,
+              address: addresses[i],
+              token: tokenInfo.symbol,
+              balance: Number(tokenJson.result) / 10 ** tokenInfo.decimals,
+              source: tokenInfo.name,
             });
           }
         }