+++ src/utils.js
@@ -27,8 +27,6 @@ export const getConfig = (env, defaultValue) => {
 /** get random item from an array. */
 export const getRandomItem = (arr) => arr[Math.floor(Math.random() * arr.length)];
 
-export const formatAddress = (address) => address && `${address.substring(0, 6)}..`;
-
 export const isAddr = data => data.length === 42;
 
 export const singleAddr = (data) => Array.isArray(data) ? data[0] : data;